+++ runtime/vue3/playground/main.ts
@@ -18,6 +18,8 @@
 
 import { createApp } from 'vue';
 
+import Core from '@tmagic/core';
+
 import App from './App.vue';
 
 Promise.all([import('../.tmagic/comp-entry'), import('../.tmagic/plugin-entry')]).then(([components, plugins]) => {
@@ -30,6 +32,19 @@ Promise.all([import('../.tmagic/comp-entry'), import('../.tmagic/plugin-entry')]
   Object.values(plugins.default).forEach((plugin: any) => {
     magicApp.use(plugin);
   });
+
+  const app = new Core({
+    ua: window.navigator.userAgent,
+    platform: 'editor',
+  });
+
+  if (app.env.isWeb) {
+    app.setDesignWidth(window.document.documentElement.getBoundingClientRect().width);
+  }
+
+  window.appInstance = app;
+  magicApp.config.globalProperties.app = app;
+  magicApp.provide('app', app);
 
   magicApp.mount('#app');
 });